genai/aws-gen-ai-kr/20_applications/08_bedrock_manus/use_cases/05_insight_extractor_strands_sdk_workshop_phase_2/requirements.txt
# uv
# ipykernel>=6.29.5
# jupyter>=1.1.1

# langchain>=0.3.26
# langgraph>=0.5.3
# strands-agents>=1.0.1
# bedrock-agentcore==0.1.1
# strands-agents-tools>=0.2.1
# mcp>=1.12.0
# requests-toolbelt>=1.0.0
# awscli>=1.41.10
# termcolor>=3.1.0
# langchain-aws>=0.2.29
# bedrock-agentcore-starter-toolkit
# uvicorn
# opentelemetry-instrumentation-langchain

ipykernel>=6.29.5
jupyter>=1.1.1
boto3>=1.40.10
strands-agents>=1.7.0
strands-agents-tools>=0.2.6
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit>=0.1.8
mcp>=1.13.0
streamlit==1.48.1
requests-toolbelt>=1.0.0
sagemaker>=2.251.1
awscli>=1.42.10
termcolor>=3.1.0
aws-opentelemetry-distro==0.12.0
matplotlib>=3.10.5
seaborn>=0.13.2
python-dotenv>=1.1.1
langchain>=0.3.27
lovelyplots>=1.0.2
weasyprint>=66.0
scikit-learn>=1.7.1
numpy>=1.26.4
pandas>=2.3.1
koreanize-matplotlib>=0.1.1
pillow>=11.3.0
plotly>=6.3.0